[test](index compaction)Add exception cases for index compaction 1. Add more exception cases for index compaction 2. Refactor index compaction test code, add a util class for all cases 3. Merge index_compaction_delete case to index compaction test
